Pundit installed, policies and tests written for User and Bicycle

+++ app/controllers/api/v1/users_controller.rb
@@ -1,12 +1,14 @@
 class Api::V1::UsersController < ApplicationController
-  before_action :set_user, only: %i[ show update destroy ]
+  before_action :set_user, only: %i[show update destroy]
+  skip_before_action :authenticate_user!, only: :create
 
   def index
-    @users = User.includes(bicycles: :chain)
+    @users = policy_scope(User.includes(bicycles: :chain))
     render json: @users, each_serializer: ::Api::V1::UserSerializer, scope: :dashboard
   end
 
   def show
+    authorize @user
     render json: @user, serializer: ::Api::V1::UserSerializer, scope: :dashboard
   end
 
@@ -14,21 +16,23 @@ class Api::V1::UsersController < ApplicationController
     @user = User.new(user_params)
 
     if @user.save
-      render json: @user, status: :created, location: @user
+      render json: @user, status: :created, serializer: ::Api::V1::UserSerializer
     else
       render json: @user.errors, status: :unprocessable_entity
     end
   end
 
   def update
+    authorize @user
     if @user.update(user_params)
-      render json: @user
+      render json: @user, serializer: ::Api::V1::UserSerializer
     else
       render json: @user.errors, status: :unprocessable_entity
     end
   end
 
   def destroy
+    authorize @user
     if @user.destroy
       head :no_content
     else
@@ -39,7 +43,7 @@ class Api::V1::UsersController < ApplicationController
   private
 
   def set_user
-    @user = User.includes(:bicycles).find(params[:id])
+    @user = User.includes(bicycles: :chain).find(params[:id])
   end
 
   def user_params
